requirements.txt
filelock==3.0.12
gunicorn==20.0.4
Fiona==1.8.4
flake8==3.8.3
Flask==1.1.2
flask-restplus==0.13.0
future==0.18.2
h5py==2.10.0
imageio 
imagesize==1.2.0
imbalanced-learn==0.5.0
imblearn==0.0
ipykernel==5.3.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.5
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications 
Keras-Preprocessing==1.1.0
keyring 
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
LBExporters==0.1.1
libarchive-c==2.9
Markdown 
MarkupSafe==1.1.1
matplotlib 
mccabe==0.6.1
networkx 
numpy==1.19.1
numpydoc 
pandas 
pandocfilters==1.4.2
pathtools==0.1.2
pexpect
pickleshare==0.7.5
Pillow 
requests 
retrying==1.3.3
rope==0.17.0
rsa==4.5
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn 
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
tensorflow==1.15.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
ujson==1.35
urllib3==1.25.9
validators==0.18.0
virtualenv==20.0.10
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0